### 2024-02-28 #227 @AndreasKuhn-ak
There are no updates one the paper. I fixed some typos in the course material in the mean time, but that should already be on github.--Sent from my Android phone with GMX Mail. Please excuse my brevity.On 2/26/24, 00:10 "Lorena A. Barba" ***@***.***> wrote:

  @AndreasKuhn-ak 👋 Let me start with an apology: I am the editor-in-chief and I have neglected JOSE for several months due to plain overwhelm. I have cut down on other service and am back here with intention of reviving our little alt-journal. Thank you for your submission! Given the time that has passed, do you have updates to the paper?
