### 2024-02-29 #227 @arm61
Hey @AndreasKuhn-ak, welcome to JOSE. Is there anyone that you could recommend, either from your knowledge of Julia people or from [this list](https://bit.ly/jose-reviewers), as a reviewer?
