### 2024-03-05 #227 @AndreasKuhn-ak
I just reviewed the list of potential reviewers, but unfortunately, I don't know any of them well enough to make a recommendation. However, I suggest (if not already done) reaching out to those who have listed Julia as one of their preferred languages. This could include individuals such as gvdr, satish-annigeri, ShahneRodgers, leios, matbesancon, davis68, gcdeshpande, HaoZeke, ma-sadeghi, and ali-ramadhan.
