### 2024-04-04 #242 @editorialbot
[REVIEW]: Scientific programming in Julia - An introductory course
**Submitting author:** @AndreasKuhn-ak (<a href="http://orcid.org/0000-0001-8146-2095">Andreas Kuhn</a>)
**Repository:** https://github.com/AndreasKuhn-ak/WS2022_Julia
**Branch with paper.md** (empty if default branch):
**Version:** V1.0.0
