### 2024-04-04 #242 @arm61
Okay, we are ready to roll!

@jarvist and @gcdeshpande, thanks for agreeing to review this exciting work! If you work through the checklist and there are any problems/comments about the material, I recommend opening an issue on [the material repository](https://github.com/AndreasKuhn-ak/WS2022_Julia), and the authors can sort them out. More information about the review guidelines can be found on the Open Journals documentation pages: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html

If anyone has any questions, ping me on here!
