### 2024-04-04 #242 @AndreasKuhn-ak
> It looks to me that the license for the code is MIT and the license for the material is CC-SA-4.0. Can you confirm this @AndreasKuhn-ak?

Yes, that is correct.
