### 2024-07-26 #242 @jarvist
Both the Release and the Paper are written targeting the Version 1.0 (Aug 2023) release. This is fine, but you may also want to update both as it's been a while & there are some edits made in the Autumn of 2023 @AndreasKuhn-ak .
