### 2024-07-26 #242 @jarvist
In terms of the paper, I think one thing that would be useful to add is some idea of what level of programming background is ideal. You say that any scientific programmer can use it, but would the content still be useful for an intermediate or further on programmer? Similarly, in the 'experience of use' section it would be good to know what the background of the 13 self-study people were. PhD students? Undergraduates? People in industry? etc.
