### 2024-07-26 #242 @jarvist
Also, is the intent for this material to be a self-study only course, or do you envisage that it would be useful / adaptable for classroom teaching? Some kind of more clear guide in the documentation would be really useful.
