### 2024-08-21 #242 @AndreasKuhn-ak
> Both the Release and the Paper are written targeting the Version 1.0 (Aug 2023) release. This is fine, but you may also want to update both as it's been a while & there are some edits made in the Autumn of 2023 @AndreasKuhn-ak .

Sorry for the late response. I was on vacation when you wrote, and I didn't check my emails thoroughly afterward. I can create a new release that incorporates the changes I made in Autumn 2023. I don't think the paper needs an update regarding the new release, as these changes were only corrections of some typos.
