### 2024-08-21 #242 @AndreasKuhn-ak
> Also, is the intent for this material to be a self-study only course, or do you envisage that it would be useful / adaptable for classroom teaching? Some kind of more clear guide in the documentation would be really useful.

I would say the primary intent is for this to be a self-study course, and we have used it in this way. However, I believe the course can also be effectively taught in a classroom setting for undergraduate students of any subject and graduate students without prior experience in programming, without any modifications. For example, a suitable format could be a block course where, in the morning, there is a teaching block of one or two lessons (depending on the size), and in the afternoon, students work alone or in groups on the exercises. In the evening or the next day, there could be a session where the solutions to the exercises are presented, and any questions regarding the exercises are answered.

I am very confident that this approach would work without major problems, as we teach a very similar Python course in exactly this format every semester for undergraduate biology students.

I will expand the target audience section accordingly.

@jarvist Thank you for taking the time to review our paper.
