### 2024-08-21 #242 @AndreasKuhn-ak
@jarvist,
Thank you for your suggestions. I have updated the paper and README file based on your feedback.
I believe it would be best to only create a new release once the review process is complete. This will ensure that the final, polished version is shared, rather than potentially releasing multiple updates during the review stage.
Please let me know if you have any other suggestions or feedback. I appreciate you taking the time to review the materials and help me improve the quality of the work.
