### 2024-08-27 #242 @arm61
> @editorialbot generate my checklist

I am investigating why this didn't work. Sorry, I didn't notice.
