### 2024-10-10 #242 @labarba
@gcdeshpande – I've edited [your comment](https://github.com/openjournals/jose-reviews/issues/242#issuecomment-2253275332) manually and copied over a review checklist for you. I don't know why it failed when you issued the command, but we should be ready to proceed now. Thank you for your contributions!
