### 2024-12-03 #242 @arm61
Hey @mlarkin863, if you comment:

> @editorialbot generate my checklist

You should get a fresh review checklist to work through. More information about reviewing for JOSE can be found here (https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html) and if you have any questions please just ask.

Thanks again for agreeing to review for JOSE!
