### 2024-12-11 #242 @mlarkin863
@AndreasKuhn-ak Hi Andreas, just as a quick thing, the link to documentation about the `IJulia` package mentioned in the README file seems to be broken?
