### 2024-12-11 #242 @AndreasKuhn-ak
@mlarkin863 Hi Matt, your right, the link is broken. I replaced it with a working link to a short tutorial on how to set up IJulia.
