### 2025-02-07 #242 @mlarkin863
@AndreasKuhn-ak Hi Andreas, apologies for the wait on an update. Just a few notes on the JOSE paper, primarily formatting. Content is all really great from reading through :)

For the PDF draft of the JOSE paper:
- In the references section is it the journal style to not have numbers indexed next to each reference item?
- In your "statement of need" section your references are cited using "surname, year" format, whereas in the "Target audience" section you cite references "1,2,3" numerically.
- Reference "Kuhn, A. (2022)" which points to your medium article i can see that the link address is a bit too long to fit cleanly within the text margins. I take it that there is no workaround for this?
- In the footer of the document there seems to be a formatting error with "¿PAGE?"
