### 2025-02-11 #242 @AndreasKuhn-ak
@mlarkin863 Thank you for your feedback. Regarding your points:

- As you suspected, this is indeed the journal's style. I used the provided GitHub Actions PDF generator that builds the PDF from the paper.md file in the repository (https://github.com/marketplace/actions/open-journals-pdf-generator)

- Yes, that was a deliberate choice since these three references link to websites providing Julia courses. They are not traditional academic publications - for instance, the first course doesn't clearly state its author. The last link points to JuliaAcademy, which hosts multiple courses. I can replace these with conventional references if needed

- I've fixed that issue.

- This appears to be because the paper is not yet published, so there is no existing page to link to.
