### 2025-02-12 #242 @mlarkin863
Hi @AndreasKuhn-ak, here are some comments from going through the full lesson plan. Apologies for the delay. All the content looks really good and the way you've structured it is super easy to follow. Can see why the student's who you have shared it with have found it useful. Very much like the different level in exercise difficulty.

Most of these are just some typos and things that I've notices but are as follows:
- "Part_3_Datatypes" -> "5. Dict - Dictionaries:" In line / cell 89 the comment states "# tuple as key, tuple as value", I think it should be "# tuple as key, string as value" Also, change "3. Datatypes" to "Datatypes" in cell 2. All other notebooks don't have a numerical index  for the notebook title (or alternatively, add numerical indexed to all titles, as long as consistent).
- "Part_5_Functions" -> "1.1. rand()" In cell 29 titled "Comments" , the following line has a type  "Randomly generated numbers are used to model this ~noisy~ **_noise_** in simulations." (Also the "1.1." is the only section number that has a "." at the end, the rest are all "1.x" or 1.x.y").
- "Part_5_Functions" -> "1.9 Keyword arguments" , in cell 75 "~togehter~ **_together"_** is spelt wrong.
- "Part_7_File_Management" -> "2.2.1 Export Data". Should there maybe be another section starting before cell 52 called "2.2.2 Import Data? Also a typo in cell 63 "2. Transform the `String` into ~a~ **_an_** array of `Strings` ...".
-  "Part_8_Data_Analysis" -> "3. Cleaning and structuring data" typos in cell 27 "Let's add columns for year, month, day and a ~combines colum~ **_combined column_** for days and month. This can be done using the ~function~ **_functions_** `year()`, `month()`, `day()` and `monthday()` ..." (separate out the code blocks)
- "Part_9_Example_Project" Same issue as first bullet point. Either remove the numerical index from cell 2 or change all the other notebooks. Also, all subsequent sections start from an numerical index of "10.x", which I assume should be "1.x"
