### 2025-02-13 #242 @AndreasKuhn-ak
@mlarkin863 Thank you for your time and thorough examination of the course. I have fixed all the mentioned typos and titles/headings in the latest commit to the repository.
