### 2025-11-28 #242 @arm61
@AndreasKuhn-ak, I am taking on the reviewing of this work. I have opened an issue on the resource repo, as the binder instance seems not to work.
