### 2025-12-03 #242 @AndreasKuhn-ak
@arm61 Thanks for pushing this forward. I am quite busy at the moment, but I will try to look at it at the end of the week or next week.
