### 2026-01-08 #242 @AndreasKuhn-ak
Sorry for the late response. I have updated the course, binder should now be working.
